ai_engine/requirements.txt
google-genai numpy requests python-dotenv